move files to app/main/java/chess
bit of a sock drawer for now
