+++ app/src/main/java/chess/GameCompletionState.java
@@ -1,4 +1,4 @@
-package game;
+package chess;
 
 
 public enum GameCompletionState {
