+++ app/src/main/java/chess/PlayerRecord.java
@@ -1,4 +1,4 @@
-package game;
+package chess;
 
 public class PlayerRecord extends PlayerInfo {
