+++ app/src/main/java/chess/Position.java
@@ -1,4 +1,4 @@
-package game;
+package chess;
 
 public class Position {
     File file;
